# improving regex speed 2x faster