Instantiating regex is expensive.  Should always keep it global for speed.  This PR improves performance over 2x times.

Original speed:

    BenchmarkPath			   10000	    106790 ns/op
    BenchmarkName			   20000	     62479 ns/op
    BenchmarkHTMLAllowed	    2000	   1149116 ns/op

Improvements in this pull request:

    BenchmarkPath			   30000	     51607 ns/op
    BenchmarkName			  100000	     23429 ns/op
    BenchmarkHTMLAllowed	    3000	    501696 ns/op

Benchmarked on Go1.4.1