@eduncan911 2015-02-20
FYI, I reverted the changes to HTMLAllowed so it's speed is back up to 1149116 as it was originally.

To speed that up, one should move the string arrays out to be global vars - so garbage collection doesn't keep getting beat on.

if you want, i can do that too and amend this PR.