@eduncan911 2015-02-20
since u have merged the other one (for docs), I can resubmit this PR with the conflicts resolved if u like?