@eduncan911 2015-02-20
there ya go.  pulled in the doc updates and resolved the merge