@kennygrant 2015-02-22
Other pull request was merged because it is trivial, I'd like some time to think about this change.