@eduncan911 2015-02-23
no problem.  all i did was move them out of local scope and into the package scope.  i also added benchmark tests for the affected methods.

all existing tests passed, indicating that no logic was changed or broken.  i also verified there was a test for each method affected by this change.

MustCompile is an extra safety measure ensuring the regex is valid before the app runs.