@kennygrant 2015-02-24
Thanks for this. I've merged your changes with minor tweaks - I prefer to keep the pkg globals inline above the functions which use them, so that it's clear where they are used, and I wanted the global names a bit clearer.